Huge changes and big improvements
Separated wrapper and bot into two files. Now bot is a class and can be
imported using `from bot import CatGirlBot` and with your own scripts.
Now you need to send `/NAME` instead of `NAME` to trigger corresponding
event - does not apply to variants. Config file is now in YAML format,
an example will be uploaded in the next commit. Added license file and
copywrite notices to the code.